fix(front): restructure routing to map the new mentorship approach
