--- redi-connect-front/src/pages/app/dashboard/mentee/CurrentMentorProfile.tsx
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export const CurrentMentorProfile = () => <h3>Current mentor profile</h3>
