--- redi-connect-front/src/pages/app/dashboard/mentor/DashboardMentor.tsx
@@ -1,4 +1,0 @@
-import React from 'react'
-import Applications from '../../applications/Applications'
-
-export const DashboardMentor = () => <Applications />
