+++ redi-connect-front/src/pages/app/find-a-mentor/FindAMentor.scss
@@ -1,5 +1,5 @@
 @import "~bulma/sass/utilities/_all";
-@import "../../../../styles/_variables.scss";
+@import "../../../styles/_variables.scss";
 
 .filters {
   margin-bottom: 1.5rem;
